refactor: 기존 컴포넌트 재사용
